allow parallel execution of userscript

+++ pakkujs/core/userscript.ts
@@ -11,10 +11,10 @@ export interface UserscriptEnv {
 }
 
 type ArgType = (
-    {type: 'init'}
-    | {type: 'pakku_before', chunk: DanmuChunk<DanmuObject>, env: UserscriptEnv}
-    | {type: 'pakku_after', chunk: DanmuChunk<DanmuObjectRepresentative>, env: UserscriptEnv}
-    | {type: 'proto_view', view: AnyObject, env: UserscriptEnv}
+    {type: 'init', env_base: UserscriptEnv | null}
+    | {type: 'pakku_before', chunk: DanmuChunk<DanmuObject>, env: Partial<UserscriptEnv>}
+    | {type: 'pakku_after', chunk: DanmuChunk<DanmuObjectRepresentative>, env: Partial<UserscriptEnv>}
+    | {type: 'proto_view', view: AnyObject, env: Partial<UserscriptEnv>}
 );
 type RetType = AnyObject;
 
@@ -23,12 +23,10 @@ const USERSCRIPT_TEMPLATE: string = userscript_template;
 export class UserscriptWorker {
     script: string;
     worker: Worker;
-    init_error: null | ErrorEvent = null;
     terminated: boolean;
 
-    resolve: null | ((res: RetType)=>void);
-    reject: null | ((e: any)=>void);
-    queue: Array<[ArgType, ((res: RetType)=>void), ((e: any)=>void)]>;
+    queue_serial: int;
+    queue_callback: Map<int, [((res: RetType)=>void), ((e: any)=>void)]>;
 
     n_before: int;
     n_after: int;
@@ -40,61 +38,41 @@ export class UserscriptWorker {
             USERSCRIPT_TEMPLATE.replace('/* MAIN */', this.script+'\n'),
         ], {type: 'text/javascript'})));
         this.terminated = false;
-        this.resolve = null;
-        this.reject = null;
-        this.queue = [];
+        this.queue_callback = new Map();
+        this.queue_serial = 0;
         this.n_before = 0;
         this.n_after = 0;
         this.n_view = 0;
 
         this.worker.onerror = (e: ErrorEvent) => {
-            console.error('pakku userscript: INIT ERROR', e);
-            this.init_error = e;
-            if(this.reject)
-                this.reject(e);
+            console.error('pakku userscript: UNCAUGHT ERROR', e);
         }
         this.worker.onmessage = (e) => {
-            if(this.resolve && this.reject) {
-                if(e.data.error)
-                    this.reject(e.data.error);
-                else
-                    this.resolve(e.data.output);
-
-                this.resolve = null;
-                this.reject = null;
-            } else {
-                console.error('pakku userscript: BAD MESSAGE', e);
-            }
-
-            this._try_perform_work();
-        };
-    }
-
-    _try_perform_work() {
-        if(this.queue.length===0) {
-            if(this.terminated)
-                this.worker.terminate();
-            return;
-        }
-
-        if(this.resolve===null) { // idle
-            let [arg, resolve, reject] = this.queue.shift()!;
-            if(this.init_error) {
-                reject(this.init_error);
+            let serial = e.data.serial;
+            if(!this.queue_callback.has(serial)) {
+                console.error('pakku userscript: BAD SERIAL', e);
                 return;
             }
 
-            this.resolve = resolve;
-            this.reject = reject;
-            this.worker.postMessage(arg);
-        }
+            let [resolve, reject] = this.queue_callback.get(serial)!;
+            this.queue_callback.delete(serial);
+
+            if(this.terminated && this.queue_callback.size===0)
+                this.worker.terminate();
+
+            if(e.data.error)
+                reject(e.data.error);
+            else
+                resolve(e.data.output);
+        };
     }
 
     exec(arg: ArgType): Promise<RetType> {
         console.log('pakku userscript: exec', arg.type);
         return new Promise((resolve: (res: RetType)=>void, reject: (e: any)=>void) => {
-            this.queue.push([arg, resolve, reject]);
-            this._try_perform_work();
+            let serial = ++this.queue_serial;
+            this.worker.postMessage([serial, arg]);
+            this.queue_callback.set(serial, [resolve, reject]);
         });
     }
 
@@ -104,13 +82,15 @@ export class UserscriptWorker {
 
         this.terminated = true;
 
-        if(this.queue.length===0)
+        // make sure to terminate only when all tasks are done to avoid missing view responses
+
+        if(this.queue_callback.size===0)
             this.worker.terminate();
-        // else terminate until all tasks are done in _try_perform_work
+        // else: terminate until all tasks are done in this.worker.onmessage
     }
 
-    async init(): Promise<int> {
-        let {n_before, n_after, n_view} = await this.exec({type: 'init'});
+    async init(env_base: UserscriptEnv | null): Promise<int> {
+        let {n_before, n_after, n_view} = await this.exec({type: 'init', env_base: env_base});
         this.n_before = n_before;
         this.n_after = n_after;
         this.n_view = n_view;
